+++ lib/Communications/ComsManager.h
@@ -11,7 +11,6 @@
 #include <ArduinoJson.h>
 #include "Strings.h"
 #include "arduino_secrets.h"
-//#include "my_secrets.h"
 
 // ----- Telemetry & Telecommand-----
 extern volatile bool sendTelemetry;
@@ -24,7 +23,7 @@ public:
     ~ComsManager();
     void initializeComs(bool isEduroam = false);
     void sendTelemetryData(String telemetryJson);
-    int getWiFiRSSI(bool isEduroam = false);
+    float getWiFiRSSI(bool isEduroam = false);
     float getLoRaRSSI();
     float getLoRaSNR();
     void serialWifiComunication();
